* Phase6Sprint7:
  Minor updates for bug#s 1476 & 1478.
  Excluded manual CHOs from report (ER to raise bug).
  Fixed bug#1475.
  Fixed bug#1476.
  Minor modifications to penalty charge validation and error message handling.
  Message modified in 'Adjust Penalty Charges' alert panel + accessibility to this panel allowed in claim status 'InvoiceDataCalculationIncorrect'.
  Stop automatic penalty charges if manually modified (by updating the penalty alert quantity flag) if needed.
  Fixed bug#1467 - Penalty charge config box - incorrect date format error message
  bug#1428 fixed + tests updated.
  Outstanding bugs fixed.
  Outstanding bugs fixed.
  Removed method which updates the claim handler dropdown (it was called
  Minor dev bug updates.
  Rules switched-off for subscriber claims.
  Penalty charge message restricted to CHOs.
  Added state change check for penalty charge revert message.
  Added update of version number in session when caim updated.
  Fixed bug#1417 - user management company name drop down alignment error.
  Additional admin fee BRE rule check added at request of ER.
  Updated total to pay on % re-selection.

+++ idas.chox.web/src/main/java/idas/chox/web/actions/AttachmentAction.java
@@ -157,21 +157,35 @@ public class AttachmentAction extends ClaimModelAction<Attachment> {
     }
 
     public String deleteAttachment() {
+
         try {
-            LOG.debug("Deleting attachment...");
+            if (model == null) {
+                getActionResponse().AssignMessageResult("Unknown error occured trying to delete the attachment.");
+                LOG.error("Cannot delete attachment - no model");
+                return ERROR;
+            }
+
+            LOG.debug("Deleting attachment (model='{}')...", model.getClass());
             if (attachmentService.deleteAtatchment(getAuthenticatedUser().getId(), model.getId())) {
                 LOG.debug("Attachment deleted.");
                 this.getActionResponse().AssignMessageResult("File has been deleted");
             } else {
-                LOG.info("User {} cannot delete attchment {}", getAuthenticatedUser().getDisplayName(), model.getId());
+                LOG.info("User {} cannot delete attachment {}", getAuthenticatedUser().getDisplayName(), model.getId());
                 this.getActionResponse().AssignMessageResult("You do not have the necessary permissions to delete this attachment.");
                 setActionError("You do not have the necessary permissions to delete this attachment");                
                 return ERROR;
             }
 
         } catch (Exception ex) {
-            LOG.error("Exception thrown deleting attachment: {}", ex.getMessage());
-            this.getActionResponse().AssignMessageResult(ex.getMessage());
+            LOG.error("Exception thrown deleting attachment with userId={}, modelId={}", getAuthenticatedUser().getId(), model.getId());
+
+            if (ex != null) {
+                LOG.error("Exception thrown deleting attachment: {}", ex.getMessage());
+                this.getActionResponse().AssignMessageResult(ex.getMessage());
+            } else {
+                LOG.error("Empty Exception thrown deleting attachment!");
+                this.getActionResponse().AssignMessageResult("Unknown error occured trying to delete the attachment.");
+            }
             setActionError(formErrorMessage(ex));
             return ERROR;
         }
@@ -395,6 +409,8 @@ public class AttachmentAction extends ClaimModelAction<Attachment> {
         model.setFileBuffer(obj);
         claim.addAttachment(model);
         claimService.updateClaim(claim);
+        getSession().put("claimDetailPageClaimId", claim.getId().intValue());
+        getSession().put("claimDetailPageClaimVersion", claim.getVersion());
     }
     // </editor-fold>
